Do not create `self.Worker` in a window.
`self.Worker` needs to be created in a web worker; in particular
`self.Worker.postMessage` is needed.  It is also needed in windows that support
web workers but lack `MessageChannel`.

However, for browsers that do not even support web workers, we do not want to
create a `Worker` object as this may give false positives to feature detection.