Fixed issue with messages missing event field
Looks like in the latest Firefox MessageChannel is still being 
supplanted by this library, however because MessagePort is implemented 
libraries from frames without MessageChannel.js are pushing valid 
MessageEvents.  As a result decodeEvent is receiving a legitimate 
MessageEvent that does not need to be faked.  For these events, just 
pass them on as-is to avoid an ‘event is undefined’ error in the 
decoding.